requirements.txt
flask
requests
pillow
numpy
scikit-image
opencv-python
ultralytics
transformers
pytesseract
tesseract
optimum[onnxruntime]
